Generic policy evaluation with hardcoded query
This is an experiment to produce a generic evaluation
function for different polices, which is able to return
arbitrary JSON result after policy execution. For the
moment a lot of conventions have to be followed in order
for this to work, which is not so good. More experiments
will follow to reduce the conventions to minimum (it would
be best to remove all of them).